refactor: make gtMetadata own component
